Implemented authentication
Implemented authentication using next-auth with prisma adapter
