+++ .env.example
@@ -1,1 +1,5 @@
 DATABASE_URL="database url"
+GOOGLE_OAUTH_CLIENT_ID="google oauth client id"
+GOOGLE_OAUTH_CLIENT_SECRET="google oauth client secret"
+NEXTAUTH_URL="next auth url"
+NEXTAUTH_SECRET="next auth secret"
