+++ package.json
@@ -11,6 +11,7 @@
     "generate": "node src/dummy.js"
   },
   "dependencies": {
+    "@next-auth/prisma-adapter": "^1.0.7",
     "@prisma/client": "^5.0.0",
     "@radix-ui/react-icons": "^1.3.0",
     "axios": "^1.4.0",
