+++ prisma/schema.prisma
@@ -7,23 +7,57 @@ datasource db {
   url      = env("DATABASE_URL")
 }
 
-enum AuthProvider {
-  Google
+model User {
+  id                String    @id @default(auto()) @map("_id") @db.ObjectId
+  name              String?
+  email             String?   @unique
+  emailVerified     DateTime?
+  image             String?
+  accounts          Account[]
+  sessions          Session[]
+  bids              Bid[]
+  listedProducts    Product[] @relation(name: "listedProducts")
+  purchasedProducts Product[] @relation(name: "purchasedProducts")
+  favorites         Product[] @relation(name: "favorites", fields: [favoriteIds], references: [id])
+  favoriteIds       String[]  @db.ObjectId
+  createdAt         DateTime  @default(now())
+  updatedAt         DateTime  @updatedAt
+  Admin             Admin?
 }
 
-model User {
-  id                String       @id @default(auto()) @map("_id") @db.ObjectId
-  name              String
-  email             String       @unique
-  provider          AuthProvider
-  bids              Bid[]
-  listedProducts    Product[]    @relation(name: "listedProducts")
-  purchasedProducts Product[]    @relation(name: "purchasedProducts")
-  favorites         Product[]    @relation(name: "favorites", fields: [favoriteIds], references: [id])
-  favoriteIds       String[]     @db.ObjectId
-  createdAt         DateTime     @default(now())
-  updatedAt         DateTime     @updatedAt
-  Admin             Admin?
+model Account {
+  id                String  @id @default(auto()) @map("_id") @db.ObjectId
+  userId            String  @db.ObjectId
+  type              String
+  provider          String
+  providerAccountId String
+  refresh_token     String?
+  access_token      String?
+  expires_at        Int?
+  token_type        String?
+  scope             String?
+  id_token          String?
+  session_state     String?
+  user              User    @relation(fields: [userId], references: [id], onDelete: Cascade)
+
+  @@unique([provider, providerAccountId])
+}
+
+model Session {
+  id           String   @id @default(auto()) @map("_id") @db.ObjectId
+  sessionToken String   @unique
+  userId       String   @db.ObjectId
+  expires      DateTime
+  user         User     @relation(fields: [userId], references: [id], onDelete: Cascade)
+}
+
+model VerificationToken {
+  id         String   @id @default(auto()) @map("_id") @db.ObjectId
+  identifier String
+  token      String   @unique
+  expires    DateTime
+
+  @@unique([identifier, token])
 }
 
 model Admin {
