+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,6 @@
+import { options } from "@/lib/auth";
+import NextAuth from "next-auth";
+
+const handler = NextAuth(options);
+
+export { handler as GET, handler as POST };
