+++ src/middleware/admin.ts
@@ -1,6 +1,6 @@
-// import { getAdminByUserId } from "@/database/admin";
+import { getAdminByUserId } from "@/database/admin";
 import type { MiddlewareFactory } from "./types";
-// import { NextResponse } from "next/server";
+import { NextResponse } from "next/server";
 const adminsOnlyRoutes: string[] = [
   // regex for routes that require admin authentication
 ];
@@ -13,13 +13,13 @@ const withAdminsAuthentication: MiddlewareFactory =
       if (regex.test(path)) {
         // the route is in the list of routes that require authentication
         // validate if the user is authenticated
-        // if (request.user == null) {
-        //   return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
-        // }
-        // const admin = getAdminByUserId(request.user.id);
-        // if (admin instanceof Error) {
-        //   return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
-        // }
+        if (request.user == null) {
+          return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+        }
+        const admin = getAdminByUserId(request.user.id);
+        if (admin instanceof Error) {
+          return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+        }
         return await next(request, event);
       }
     }
